Логика DataProvider перенесена в Store-класс
